src/requirements.txt
flask
flask-restx
mysql-connector-python
pytest
